Move widget to databases and rename it
Rename `OptimadeQueryWidget` to `AiidalabOptimadeWidget`.
Use `ase.Atoms` for the `structure` traitlet instead of
`aiida.orm.StructureData`, since `CodQueryWidget` is doing the same, and
we then don't need to import from AiiDA in this file.
